requirements.txt
pymongo==4.13.0
python-dotenv==1.1.0
faker==37.1.0